+++ src/server/game/Scripting/ScriptDefines/InstanceMapScript.cpp
@@ -1,0 +1,27 @@
+/*
+ * This file is part of the AzerothCore Project. See AUTHORS file for Copyright information
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Affero General Public License as published by the
+ * Free Software Foundation; either version 3 of the License, or (at your
+ * option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "ScriptMgr.h"
+#include "ScriptMgrMacros.h"
+
+InstanceScript* ScriptMgr::CreateInstanceScript(InstanceMap* map)
+{
+    ASSERT(map);
+
+    auto tempScript = ScriptRegistry<InstanceMapScript>::GetScriptById(map->GetScriptId());
+    return tempScript ? tempScript->GetInstanceScript(map) : nullptr;
+}
